fix: host engine cancel if process is nil

+++ engine/host/cancel.go
@@ -2,6 +2,10 @@ package host
 
 // Cancel cancels the command.
 func (h *host) Cancel() error {
+	if h.cmd.Process == nil {
+		return nil
+	}
+
 	if err := h.cmd.Process.Kill(); err != nil {
 		return err
 	}
